Úprava rozhraní NumberConverter
NumberConverter.ToString(double value, int numbase, string format) na
NumberConverter.ToString(double value, NumSystem numbase, string format)
Nyní se zadává soustava pomocí enum.
Současně byla provedena úprava kódu podle nového rozhraní.

+++ src/CalculatorProject/CalcUnitTester/Program.cs
@@ -48,10 +48,10 @@ namespace CalcUnitParser
 
         Calculation calculation = new Calculation(expr);
         Console.WriteLine("Result: {0}", calculation.Value);
-        Console.WriteLine("GetAsString <Dec>: {0}", NumberConverter.ToString(calculation.Value, (int)NumSystem.Dec, ""));
-        Console.WriteLine("GetAsString <Bin>: {0}", NumberConverter.ToString(calculation.Value, (int)NumSystem.Bin, ""));
-        Console.WriteLine("GetAsString <Oct>: {0}", NumberConverter.ToString(calculation.Value, (int)NumSystem.Oct, ""));
-        Console.WriteLine("GetAsString <Hex>: {0}", NumberConverter.ToString(calculation.Value, (int)NumSystem.Hex, ""));
+        Console.WriteLine("GetAsString <Dec>: {0}", NumberConverter.ToString(calculation.Value, NumSystem.Dec, ""));
+        Console.WriteLine("GetAsString <Bin>: {0}", NumberConverter.ToString(calculation.Value, NumSystem.Bin, ""));
+        Console.WriteLine("GetAsString <Oct>: {0}", NumberConverter.ToString(calculation.Value, NumSystem.Oct, ""));
+        Console.WriteLine("GetAsString <Hex>: {0}", NumberConverter.ToString(calculation.Value, NumSystem.Hex, ""));
       } while (true);
     }
   }
